Add full roundtrip fuzzing
