+++ Makefile
@@ -23,10 +23,10 @@ lint: ## Runs the linters (including internal ones)
 build: ## Build an lbadd binary that is ready for prod
 	go build -o lbadd -ldflags="-s -w -X 'main.Version=$(shell date +%Y%m%d)'" ./cmd/lbadd
 
-.PHONY: fuzzy-parser
-fuzzy-parser: ## Starts fuzzing the parser
-	go-fuzz-build -o parser-fuzz.zip ./internal/parser
-	go-fuzz -bin parser-fuzz.zip -workdir internal/parser/test/fuzz
+.PHONY: fuzz
+fuzz: ## Starts fuzzing the database
+	go-fuzz-build -o lbadd-fuzz.zip ./internal/test
+	go-fuzz -bin lbadd-fuzz.zip -workdir internal/test/testdata/fuzz
 
 ## Help display.
 ## Pulls comments from beside commands and prints a nicely formatted
