+++ go.mod
@@ -3,6 +3,7 @@ module github.com/tomarrell/lbadd
 go 1.13
 
 require (
+	github.com/dvyukov/go-fuzz v0.0.0-20200318091601-be3528f3a813 // indirect
 	github.com/google/go-cmp v0.5.0
 	github.com/kr/text v0.2.0 // indirect
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
